tools/requirements.txt
PyYAML==6.0
stix2==2.1.0
